moving to common functions

+++ functions/fn_b5_install_motd.sh
@@ -2,21 +2,23 @@
 # install and customize the "Mot Of The Day" screen
 # ------------------------------------------------------------------------------
 
-install_motd() {
+Install.motd() {
+	local p=/etc/update-motd.d
+
 	# abort if MOTD is already installed
-	[ -s "/etc/update-motd.d/*-footer" ] && return
+	[ -s "$p/*-footer" ] && return
 
 	# verify needed packages
 	Pkg.requires figlet lsb-release
 
 	# copying files & make them executables
-	mkdir -p /etc/update-motd.d && cd "$_"
-	rm -rf ./*
-	File.into . motd/*
-	chmod +x ./*
+	mkdir -p "$p"
+	rm -rf $p/*
+	File.into $p motd/*
+	chmod +x $p/*
 
 	# remove /etc/motd on pure debian
-	rm -f /etc/motd
+	[ "$ENV_product" = "debian" ] && rm -f /etc/motd
 
 	Msg.info "Customization of MOTD completed!"
-}	# end install_motd
+}	# end Install.motd
